+++ custom_components/pioneer_async/const.py
@@ -56,7 +56,7 @@ DATA_SCHEMA = vol.Schema(
 OPTIONS_SCHEMA = vol.Schema(
     {
         vol.Optional(CONF_SCAN_INTERVAL, default=DEFAULT_SCAN_INTERVAL): int,
-        vol.Optional(CONF_TIMEOUT, default=DEFAULT_TIMEOUT): int,
+        vol.Optional(CONF_TIMEOUT, default=DEFAULT_TIMEOUT): float,
         vol.Optional(CONF_COMMAND_DELAY, default=DEFAULT_COMMAND_DELAY): float,
         vol.Optional(CONF_VOLUME_WORKAROUND, default=DEFAULT_VOLUME_WORKAROUND): bool,
     }
